tools/libxl: filter afterpoll_check_fd() events with ~POLLHUP
I haven't made this conditional with #ifdef __NetBSD__ as I believe it
may be necessary for other systems, and harmless to the rest, but it has
not yet been tested on anything other than NetBSD.

From the pkgsrc patch for the same change:

The SIGCHLD handler just writes to a pipe where the reader is the
same process. The idea is that this will cause poll(2) in the main
thread to exit with the reaper pipe readable, and do child cleanup here.

Unfortunably, is the child also has a write pipe back to the
parent (as e.g. with pygrub), the loop in afterpoll_internal() may see the
POLLHUP event on this pipe before the POLLIN even on the reaper pipe, and
this will be considered as an error (from e.g. pygrub).

work around by filtering POLLHUP events here

Signed-off-by: Greg A. Woods <woods@robohack.ca>