tools/hotplug/NetBSD: overhaul NetBSD rc.d script configuration
Change how scripts/hotplugpath.sh is used -- rc.d scripts should be
configured by a script fragment in the same-named adjacent rc.conf.d
directory, so add a line to source hotplugpath.sh to that fragment.

With proper use of scripts/hotplugpath.sh none of the rc.d scripts need
to be, nor should they be, processed by configure.  Fix the scripts to
use the variables defined in hotplughpath.sh.

Rename xen-watchdog to xenwatchdog for overall consistency.

In tools/configure.ac also make sure ../config/Paths.mk is remade in
case configure is re-run just in the tools subdir -- this should be
transparent as everything in there is supposed to be controlled by
../m4/paths.m4, which is also included in tools/configure.ac.

Note: LD_LIBRARY_PATH should probably not be needed on NetBSD, but I've
left is set for now.

Signed-off-by: Greg A. Woods <woods@robohack.ca>