tools/test/x86_emulator: improve portability of testcase generator
The od(1) program is non-standard, and the output of it's '-t x' mode
differs between the GNU utilities and those of other systems.  For one
the whitespace between columns is different.

Also both GNU od(1) and BSD od(1) may produce a final line with no data,
just a byte count.

I've fixed the 'sed' processing of the od(1) output to allow for these
differences and now both versions of od(1) can be used interchangeably.

(Note the old UNIX od(1), available up to System V Release 4, would be
unusable with no 4-byte format, nor a '-t' option)

Signed-off-by: Greg A. Woods <woods@robohack.ca>