tools/include/Makefile: fix unportable link-farm creation
It is not portable to identify directories using a trailing slash

Unix systems define the "slash" character only as a separator -- it is
not an actual component of a filename.  As a consequence any pathname,
for a file of any type, may have as many trailing slashes on the end as
one might want to append.  This has been true since the beginning of
Unix.  Some modern unix-like systems have construed that one or more
trailing slashes on a pathname might suggest that it must somehow point
to a directory, but this is not universally or historically true.

As a result the attempt to create a link-farm in this Makefile is
broken and does not work with the Unix filesystem API and a standard
POSIX shell.

Ideally it might be fixed by using find(1) to find files
only (i.e. using "-type f"), but in practice it suffices to simply
create symlinks to every filesystem object in the top-level directory,
as links to sub-directories suffice to find the files they contain.

Also the generated COPYING file is in some conflict with the one already
present in ../../xen/include/public.  The latter should be edited such
that it does not need any transformation before being installed on a
target system.  Also that existing file mentions files in "sys" which
should not be copied by the xen project in the first place, and which I
have replaced entirely elsewhere in this patchset, as per the original
pkgsrc patches, so that they don't include any foreign code anyway.)

Signed-off-by: Greg A. Woods <woods@robohack.ca>