xen/Makefile: assign to LDFLAGS instead of appending
Using the well-known LDFLAGS makefile macro with ld(1) goes against
convention and so appending to it here can cause build problems when the
build is done by another meta-build system such as pkgsrc where LDFLAGS
might be provided in the environment for use with cc(1) as is normal,
and where those values might include flags supported only by cc(1) and
not directly by ld(1), e.g. "-Wl,--no-undefined".

So, just use the already defined and meaningfully named LDFLAGS_DIRECT
variable.  We don't need or want any cc(1) LDFLAGS anywhere in here in
the first place!

Also instead of appending the dynamic LDFLAGS-y to it, just directly use
both in the final assignment to XEN_LDFLAGS.  Keep it simple, avoid
unnecessary levels of indirect use which can only result in confusion!

Signed-off-by: Greg A. Woods <woods@robohack.ca>