check-endbr.sh: hack this out for now -- not working on NetBSD!
Warning:  This patch is not intended for general use -- it is NetBSD
specific!

The check-endbr.sh script doesn't work on my NetBSD systems as-is.  This
is probably because my toolchain is too old to even generate ENDBR
instructions in the first place.

I have gcc-9.3.0 and ld 2.34.

I'm not sure how to make rules in this Makefile depend on the build OS
or toolchain versions, so for expediency I've simply disabled use of
check-endbr.sh entirely.

Signed-off-by: Greg A. Woods <woods@robohack.ca>