tools/firmware/vgabios: make the build work more portably
Herein there's some confusion about which compiler macros are used for
what purpose.

This makefile needs a preprocessor, a cross-compiler, and a build-host
compiler.

So, instead of getting them all mixed up, define and use macros for each
separately.  The preprocessor can be just "cc -E -P", so define it in
terms of the generic "$(CC)" macro, but use "$(HOSTCC)" explicitly for
the build-host compiler, and keep "$(BCC)" as the cross-compiler.

Signed-off-by: Greg A. Woods <woods@robohack.ca>