tools: improve attempts to configure and use libyajl
tools/configure.ac doesn't do all the necessary handling to find a
host-supplied libyajl, and so to build on some systems it is necessary
to use all of PREPEND_INCLUDES, PREPEND_LIB, CFLAGS, and LDFLAGS to
point to where its components are, and to use "--enable-rpath", and to
fix tools/libs/stat/Makefile to use PREPEND_LIB with RPATH as well.

Ideally libyajl should be handled the same way libuuid is handled.

Signed-off-by: Greg A. Woods <woods@robohack.ca>